Phase 2 - InstaNet/Phase 2 - Question/requirements.txt
Pillow==10.2.0
customtkinter==5.2.2
